requirements.txt
requests==2.26.0
python-dotenv==0.19.1
python-telegram-bot==11.1.0
redis==3.2.1